Add addon logs
