+++ Gopkg.toml
@@ -26,7 +26,7 @@
 
 [[constraint]]
   name = "github.com/Scalingo/go-scalingo"
-  branch = "master"
+  branch = "fix/cli/398/addon_logs"
 
 [[constraint]]
   branch = "master"
