+++ cmd/logs.go
@@ -4,6 +4,7 @@ import (
 	"github.com/Scalingo/cli/appdetect"
 	"github.com/Scalingo/cli/apps"
 	"github.com/Scalingo/cli/cmd/autocomplete"
+	"github.com/Scalingo/cli/db"
 	"github.com/urfave/cli"
 )
 
@@ -39,6 +40,38 @@ var (
 		},
 		BashComplete: func(c *cli.Context) {
 			autocomplete.CmdFlagsAutoComplete(c, "logs")
+		},
+	}
+
+	AddonsLogsCommand = cli.Command{
+		Name:     "addon-logs",
+		Category: "Addons",
+		Usage:    "Get the logs of your addons",
+		Description: `Get the logs of your addons
+   Example:
+     Get 100 lines:          'scalingo --app my-app --addon addon_uuid logs -n 100'
+     Real-Time logs:         'scalingo --app my-app --addon addon_uuid logs -f'
+		 `,
+		Flags: []cli.Flag{appFlag,
+			cli.IntFlag{Name: "lines, n", Value: 20, Usage: "Number of log lines to dump", EnvVar: ""},
+			cli.BoolFlag{Name: "follow, f", Usage: "Stream logs of addon, (as \"tail -f\")", EnvVar: ""},
+		},
+		Before: AuthenticateHook,
+		Action: func(c *cli.Context) {
+			if len(c.Args()) != 0 {
+				cli.ShowCommandHelp(c, "logs")
+				return
+			}
+			currentApp := appdetect.CurrentApp(c)
+			currentAddon := addonName(c)
+			opts := db.LogsOpts{
+				Follow: c.Bool("follow"),
+				Count:  c.Int("lines"),
+			}
+			err := db.Logs(currentApp, currentAddon, opts)
+			if err != nil {
+				errorQuit(err)
+			}
 		},
 	}
 )
