+++ logs/logs.go
@@ -1,0 +1,251 @@
+package logs
+
+import (
+	"bufio"
+	"fmt"
+	"net/url"
+	"os"
+	"os/signal"
+	"regexp"
+	"strings"
+	"time"
+	"unicode"
+
+	stdio "io"
+
+	"github.com/Scalingo/cli/config"
+	"github.com/Scalingo/cli/debug"
+	"github.com/Scalingo/cli/io"
+	"github.com/Scalingo/cli/signals"
+	"github.com/fatih/color"
+	"golang.org/x/net/websocket"
+	errgo "gopkg.in/errgo.v1"
+)
+
+type WSEvent struct {
+	Type      string    `json:"event"`
+	Log       string    `json:"log"`
+	Timestamp time.Time `json:"timestamp"`
+}
+
+func Dump(logsURL string, n int, filter string) error {
+	c := config.ScalingoClient()
+	res, err := c.Logs(logsURL, n, filter)
+	if err != nil {
+		return errgo.Mask(err, errgo.Any)
+	}
+	defer res.Body.Close()
+
+	if res.StatusCode == 404 || res.StatusCode == 204 {
+		io.Error("There is not log for this application")
+		io.Info("Ensure your application is writing to the standard output")
+		return nil
+	}
+
+	sr := bufio.NewReader(res.Body)
+	for {
+		bline, err := sr.ReadBytes('\n')
+		if err != nil {
+			break
+		}
+
+		colorizeLogs(string(bline))
+	}
+
+	return nil
+}
+
+func Stream(logsRawURL string, filter string) error {
+	var (
+		err   error
+		event WSEvent
+	)
+
+	logsURL, err := url.Parse(logsRawURL)
+	if err != nil {
+		return errgo.Mask(err, errgo.Any)
+	}
+	if logsURL.Scheme == "https" {
+		logsURL.Scheme = "wss"
+	} else {
+		logsURL.Scheme = "ws"
+	}
+
+	logsURLString := fmt.Sprintf("%s&stream=true", logsURL.String())
+	if filter != "" {
+		logsURLString = fmt.Sprintf("%s&filter=%s", logsURLString, filter)
+	}
+
+	conn, err := websocket.Dial(logsURLString, "", "http://scalingo-cli.local/"+config.Version)
+	if err != nil {
+		return errgo.Mask(err, errgo.Any)
+	}
+
+	signals.CatchQuitSignals = false
+	signals := make(chan os.Signal)
+	signal.Notify(signals, os.Interrupt)
+
+	go func() {
+		defer close(signals)
+		<-signals
+		err := conn.Close()
+		if err != nil {
+			debug.Println("Fail to close log websocket connection", err)
+		}
+	}()
+
+	for {
+		err := websocket.JSON.Receive(conn, &event)
+		if err != nil {
+			conn.Close()
+			if err == stdio.EOF {
+				debug.Println("Remote server broke the connection, reconnecting")
+				for err != nil {
+					conn, err = websocket.Dial(logsURLString, "", "http://scalingo-cli.local/"+config.Version)
+					time.Sleep(time.Second * 1)
+				}
+				continue
+
+			} else if strings.Contains(err.Error(), "use of closed network connect") {
+				return nil
+			} else {
+				return errgo.Mask(err, errgo.Any)
+			}
+		} else {
+			switch event.Type {
+			case "ping":
+			case "log":
+				colorizeLogs(strings.TrimSpace(event.Log))
+			}
+		}
+	}
+}
+
+type colorFunc func(...interface{}) string
+
+func colorizeLogs(logs string) {
+	containerColors := []colorFunc{
+		color.New(color.FgBlue).SprintFunc(),
+		color.New(color.FgCyan).SprintFunc(),
+		color.New(color.FgGreen).SprintFunc(),
+		color.New(color.FgMagenta).SprintFunc(),
+		color.New(color.FgHiYellow).SprintFunc(),
+		color.New(color.FgHiBlue).SprintFunc(),
+		color.New(color.FgHiCyan).SprintFunc(),
+		color.New(color.FgHiGreen).SprintFunc(),
+		color.New(color.FgHiMagenta).SprintFunc(),
+	}
+
+	lines := strings.Split(logs, "\n")
+
+	for _, line := range lines {
+		if line == "" {
+			continue
+		}
+
+		lineSplit := strings.Split(line, " ")
+		if len(lineSplit) < 5 {
+			fmt.Println(line)
+			continue
+		}
+		content := strings.Join(lineSplit[5:], " ")
+
+		headerSplit := lineSplit[:5]
+		date := strings.Join(headerSplit[:4], " ")
+		containerWithSurround := headerSplit[4]
+		container := containerWithSurround[1 : len(containerWithSurround)-1]
+
+		colorId := 0
+		for _, letter := range []byte(container) {
+			colorId += int(letter)
+		}
+
+		if container == "router" {
+			colorId += 6
+			content = colorizeRouterLogs(content)
+		} else {
+			content = errorHighlight(content)
+		}
+		colorId = colorId % len(containerColors)
+
+		fmt.Printf(
+			"%s [%s] %s\n",
+			color.New(color.FgYellow).Sprint(date),
+			containerColors[colorId](container),
+			content,
+		)
+	}
+}
+
+const (
+	varNameState int = iota
+	equalState
+	inTextState
+	inStringState
+)
+
+func colorizeRouterLogs(content string) string {
+	var outContent string
+	var stateBeginnedAt int
+
+	state := varNameState
+	// Remember where the matching state started
+	stateBeginnedAt = 0
+	outContent = ""
+	// Will be true if we are on the last char
+	isEnd := false
+	for i := 0; i < len([]rune(content)); i++ {
+		c := []rune(content)[i]
+		if i+1 >= len([]rune(content)) {
+			isEnd = true
+		}
+
+		// Some cases can return one char back if they go too far
+		switch state {
+		case varNameState:
+			if isEnd || (!unicode.IsLetter(c) && string(c) != "_") {
+				end := i
+				outContent += color.New(color.FgGreen).Sprint(content[stateBeginnedAt:end])
+				state = equalState
+				stateBeginnedAt = end
+				i--
+			}
+		case equalState:
+			end := i + 1
+			outContent += color.New(color.FgRed).Sprint(content[stateBeginnedAt:end])
+			state = inTextState
+			stateBeginnedAt = end
+		case inTextState:
+			if !isEnd && string(c) == "\"" {
+				state = inStringState
+			} else if isEnd || string(c) == " " {
+				end := i + 1
+				outContent += color.New(color.FgWhite).Sprint(content[stateBeginnedAt:end])
+				state = varNameState
+				stateBeginnedAt = end
+			}
+		case inStringState:
+			isEndAfter := i+2 >= len([]rune(content))
+			if !isEndAfter && string(c) == "\\" {
+				// Skip next char
+				i++
+			} else if isEnd || string(c) == "\"" {
+				state = inTextState
+				if isEnd {
+					i--
+				}
+			}
+		default:
+			outContent += string(c)
+		}
+	}
+
+	return outContent
+}
+
+func errorHighlight(content string) string {
+	reg := regexp.MustCompile("(?i)(\\berr(or)?\\b)")
+	outContent := reg.ReplaceAllString(content, color.New(color.BgRed).Sprint("$1"))
+
+	return outContent
+}
